(fix):refactor code
